reduce duplicated code required to setup the editor apps
