+++ app/ui-react/syndesis/src/modules/integrations/IntegrationCreatorApp.tsx
@@ -1,28 +1,11 @@
-/* tslint:disable:object-literal-sort-keys */
-import { getConnectionIcon, getStepIcon } from '@syndesis/api';
-import {
-  Breadcrumb,
-  IntegrationFlowStepGeneric,
-  IntegrationFlowStepWithOverview,
-  IntegrationVerticalFlow,
-} from '@syndesis/ui';
+import { Breadcrumb } from '@syndesis/ui';
 import * as React from 'react';
 import { Route, Switch } from 'react-router';
 import { Link } from 'react-router-dom';
 import { WithClosedNavigation } from '../../shared';
-import { IntegrationEditorSidebar } from './components';
 import { AddStepPage } from './components/editor/AddStepPage';
-import { ReviewPage } from './components/editor/api-provider/EditPage';
-import { EditPage } from './components/editor/api-provider/ReviewPage';
-import { UploadPage } from './components/editor/api-provider/UploadPage';
 import { EditorApp } from './components/editor/EditorApp';
-import { ConfigureActionPage } from './components/editor/endpoint/ConfigureActionPage';
-import { SelectActionPage } from './components/editor/endpoint/SelectActionPage';
 import { SaveIntegrationPage } from './components/editor/SaveIntegrationPage';
-import { SelectConnectionPage } from './components/editor/SelectConnectionPage';
-import { ConfigureStepPage } from './components/editor/step/ConfigureStepPage';
-import { TemplateStepPage } from './components/editor/template/TemplateStepPage';
-import { getDataShapeText, toUIStepKind } from './components/editor/utils';
 import resolvers from './resolvers';
 import routes from './routes';
 
@@ -77,641 +60,6 @@ const saveIntegrationPage = (
   />
 );
 
-const startStepSelectConnectionPage = (
-  <SelectConnectionPage
-    cancelHref={resolvers.list}
-    apiProviderHref={resolvers.create.start.apiProvider.upload}
-    connectionHref={(connection, params, state) =>
-      resolvers.create.start.connection.selectAction({
-        connection,
-        ...params,
-        ...state,
-      })
-    }
-    filterHref={resolvers.create.start.basicFilter}
-    mapperHref={resolvers.create.start.dataMapper}
-    templateHref={resolvers.create.start.template}
-    stepHref={resolvers.create.start.step}
-    sidebar={() => (
-      <IntegrationVerticalFlow>
-        {({ expanded }) => (
-          <>
-            <IntegrationFlowStepGeneric
-              icon={<i className={'fa fa-plus'} />}
-              i18nTitle={'1. Start'}
-              i18nTooltip={'Start'}
-              active={true}
-              showDetails={expanded}
-              description={'Choose a connection'}
-            />
-            <IntegrationFlowStepWithOverview
-              icon={<i className={'fa fa-plus'} />}
-              i18nTitle={'2. Finish'}
-              i18nTooltip={'Finish'}
-              active={false}
-              showDetails={expanded}
-              name={'n/a'}
-              action={'n/a'}
-              dataType={'n/a'}
-            />
-          </>
-        )}
-      </IntegrationVerticalFlow>
-    )}
-  />
-);
-
-const startStepSelectActionPage = (
-  <SelectActionPage
-    cancelHref={resolvers.create.start.selectStep}
-    sidebar={({ connection }) => (
-      <IntegrationVerticalFlow>
-        {({ expanded }) => (
-          <>
-            <IntegrationFlowStepGeneric
-              icon={
-                <img
-                  alt={'Step icon'}
-                  src={getConnectionIcon(process.env.PUBLIC_URL, connection)}
-                  width={24}
-                  height={24}
-                />
-              }
-              i18nTitle={`1. ${connection.name}`}
-              i18nTooltip={`1. ${connection.name}`}
-              active={true}
-              showDetails={expanded}
-              description={'Choose an action'}
-            />
-            <IntegrationFlowStepWithOverview
-              icon={<i className={'fa fa-plus'} />}
-              i18nTitle={'2. Finish'}
-              i18nTooltip={'Finish'}
-              active={false}
-              showDetails={expanded}
-              name={'n/a'}
-              action={'n/a'}
-              dataType={'n/a'}
-            />
-          </>
-        )}
-      </IntegrationVerticalFlow>
-    )}
-    selectHref={(actionId, p, s) =>
-      resolvers.create.start.connection.configureAction({
-        actionId,
-        ...p,
-        ...s,
-      })
-    }
-  />
-);
-
-const startStepConfigureActionPage = (
-  <ConfigureActionPage
-    backHref={(p, s) =>
-      resolvers.create.start.connection.selectAction({ ...p, ...s })
-    }
-    cancelHref={resolvers.list}
-    mode={'adding'}
-    nextStepHref={(p, s) =>
-      resolvers.create.start.connection.configureAction({
-        ...p,
-        ...s,
-      })
-    }
-    sidebar={({ connection }) => (
-      <IntegrationVerticalFlow>
-        {({ expanded }) => (
-          <>
-            <IntegrationFlowStepGeneric
-              icon={
-                <img
-                  alt={'Step icon'}
-                  src={getConnectionIcon(process.env.PUBLIC_URL, connection)}
-                  width={24}
-                  height={24}
-                />
-              }
-              i18nTitle={`1. ${connection.connector!.name}`}
-              i18nTooltip={`1. ${connection.name}`}
-              active={true}
-              showDetails={expanded}
-              description={'Configure the action'}
-            />
-            <IntegrationFlowStepWithOverview
-              icon={<i className={'fa fa-plus'} />}
-              i18nTitle={'2. Finish'}
-              i18nTooltip={'Finish'}
-              active={false}
-              showDetails={expanded}
-              name={'n/a'}
-              action={'n/a'}
-              dataType={'n/a'}
-            />
-          </>
-        )}
-      </IntegrationVerticalFlow>
-    )}
-    postConfigureHref={(integration, params, state) => {
-      return resolvers.create.finish.selectStep({
-        integration,
-        ...params,
-        position: '1',
-      });
-    }}
-  />
-);
-
-const startStepConfigureStepPage = (
-  <ConfigureStepPage
-    cancelHref={resolvers.list}
-    mode={'adding'}
-    sidebar={({ step }) => (
-      <IntegrationVerticalFlow>
-        {({ expanded }) => {
-          return (
-            <>
-              <IntegrationFlowStepGeneric
-                icon={
-                  <img
-                    alt={'Step icon'}
-                    src={getStepIcon(process.env.PUBLIC_URL, step)}
-                    width={24}
-                    height={24}
-                  />
-                }
-                i18nTitle={`1. ${step.name}`}
-                i18nTooltip={`1. ${step.description}`}
-                active={true}
-                showDetails={expanded}
-                description={'Configure the action'}
-              />
-              <IntegrationFlowStepWithOverview
-                icon={<i className={'fa fa-plus'} />}
-                i18nTitle={'2. Finish'}
-                i18nTooltip={'Finish'}
-                active={false}
-                showDetails={expanded}
-                name={'n/a'}
-                action={'n/a'}
-                dataType={'n/a'}
-              />
-            </>
-          );
-        }}
-      </IntegrationVerticalFlow>
-    )}
-    postConfigureHref={(integration, params) =>
-      resolvers.create.finish.selectStep({
-        integration,
-        ...params,
-      })
-    }
-  />
-);
-
-const finishStepSelectConnectionPage = (
-  <SelectConnectionPage
-    cancelHref={resolvers.list}
-    apiProviderHref={resolvers.create.finish.apiProvider.upload}
-    connectionHref={(connection, params, state) =>
-      resolvers.create.finish.connection.selectAction({
-        connection,
-        ...params,
-        ...state,
-      })
-    }
-    filterHref={resolvers.create.finish.basicFilter}
-    mapperHref={resolvers.create.finish.dataMapper}
-    templateHref={resolvers.create.finish.template}
-    stepHref={(step, params, state) =>
-      resolvers.create.finish.step({
-        step,
-        ...params,
-        ...state,
-      })
-    }
-    sidebar={({ steps }) => (
-      <IntegrationVerticalFlow>
-        {({ expanded }) => {
-          const s = toUIStepKind(steps[0]);
-          return (
-            <>
-              <IntegrationFlowStepWithOverview
-                icon={<img alt={s.name} src={s.icon} width={24} height={24} />}
-                i18nTitle={`1. ${s.name}`}
-                i18nTooltip={`1. ${s.title}`}
-                active={false}
-                showDetails={expanded}
-                name={s.name}
-                action={s.action && s.action.name}
-                dataType={getDataShapeText(s.stepKind!, s.outputDataShape)}
-              />
-              <IntegrationFlowStepGeneric
-                icon={<i className={'fa fa-plus'} />}
-                i18nTitle={'2. Finish'}
-                i18nTooltip={'Finish'}
-                active={true}
-                showDetails={expanded}
-                description={'Choose a connection'}
-              />
-            </>
-          );
-        }}
-      </IntegrationVerticalFlow>
-    )}
-  />
-);
-
-const finishStepSelectActionPage = (
-  <SelectActionPage
-    cancelHref={(p, s) => resolvers.create.finish.selectStep({ ...p, ...s })}
-    sidebar={({ connection, steps }) => (
-      <IntegrationVerticalFlow>
-        {({ expanded }) => {
-          const s = toUIStepKind(steps[0]);
-          return (
-            <>
-              <IntegrationFlowStepWithOverview
-                icon={<img alt={s.name} src={s.icon} width={24} height={24} />}
-                i18nTitle={`1. ${s.name}`}
-                i18nTooltip={`1. ${s.title}`}
-                active={false}
-                showDetails={expanded}
-                name={s.name}
-                action={s.action && s.action.name}
-                dataType={getDataShapeText(s.stepKind!, s.outputDataShape)}
-              />
-              <IntegrationFlowStepGeneric
-                icon={
-                  <img
-                    alt={'Step icon'}
-                    src={getConnectionIcon(process.env.PUBLIC_URL, connection)}
-                    width={24}
-                    height={24}
-                  />
-                }
-                i18nTitle={`2. ${connection.connector!.name}`}
-                i18nTooltip={`2. ${connection.name}`}
-                active={true}
-                showDetails={expanded}
-                description={'Choose an action'}
-              />
-            </>
-          );
-        }}
-      </IntegrationVerticalFlow>
-    )}
-    selectHref={(actionId, p, s) =>
-      resolvers.create.finish.connection.configureAction({
-        actionId,
-        ...p,
-        ...s,
-      })
-    }
-  />
-);
-
-const finishStepConfigureActionPage = (
-  <ConfigureActionPage
-    backHref={(p, s) =>
-      resolvers.create.finish.connection.selectAction({ ...p, ...s })
-    }
-    cancelHref={resolvers.list}
-    mode={'adding'}
-    nextStepHref={(p, s) =>
-      resolvers.create.finish.connection.configureAction({
-        ...p,
-        ...s,
-      })
-    }
-    sidebar={({ connection, steps }) => (
-      <IntegrationVerticalFlow>
-        {({ expanded }) => {
-          const s = toUIStepKind(steps[0]);
-          return (
-            <>
-              <IntegrationFlowStepWithOverview
-                icon={<img alt={s.name} src={s.icon} width={24} height={24} />}
-                i18nTitle={`1. ${s.name}`}
-                i18nTooltip={`1. ${s.title}`}
-                active={false}
-                showDetails={expanded}
-                name={s.name}
-                action={s.action && s.action.name}
-                dataType={getDataShapeText(s.stepKind!, s.outputDataShape)}
-              />
-              <IntegrationFlowStepGeneric
-                icon={
-                  <img
-                    alt={'Step icon'}
-                    src={getConnectionIcon(process.env.PUBLIC_URL, connection)}
-                    width={24}
-                    height={24}
-                  />
-                }
-                i18nTitle={`2. ${connection.connector!.name}`}
-                i18nTooltip={`2. ${connection.name}`}
-                active={true}
-                showDetails={expanded}
-                description={'Configure the action'}
-              />
-            </>
-          );
-        }}
-      </IntegrationVerticalFlow>
-    )}
-    postConfigureHref={(integration, params) =>
-      resolvers.create.configure.index({
-        integration,
-        ...params,
-      })
-    }
-  />
-);
-
-const finishStepConfigureStepPage = (
-  <ConfigureStepPage
-    cancelHref={resolvers.list}
-    mode={'adding'}
-    sidebar={({ step, steps }) => (
-      <IntegrationVerticalFlow>
-        {({ expanded }) => {
-          const s = toUIStepKind(steps[0]);
-          return (
-            <>
-              <IntegrationFlowStepWithOverview
-                icon={<img alt={s.name} src={s.icon} width={24} height={24} />}
-                i18nTitle={`1. ${s.name}`}
-                i18nTooltip={`1. ${s.title}`}
-                active={false}
-                showDetails={expanded}
-                name={s.name}
-                action={s.action && s.action.name}
-                dataType={getDataShapeText(s.stepKind!, s.outputDataShape)}
-              />
-              <IntegrationFlowStepGeneric
-                icon={
-                  <img
-                    alt={'Step icon'}
-                    src={getStepIcon(process.env.PUBLIC_URL, step)}
-                    width={24}
-                    height={24}
-                  />
-                }
-                i18nTitle={`2. ${step.name}`}
-                i18nTooltip={`2. ${step.description}`}
-                active={true}
-                showDetails={expanded}
-                description={'Configure the action'}
-              />
-            </>
-          );
-        }}
-      </IntegrationVerticalFlow>
-    )}
-    postConfigureHref={(integration, params) =>
-      resolvers.create.configure.index({
-        integration,
-        ...params,
-      })
-    }
-  />
-);
-
-const addStepSelectConnectionPage = (
-  <SelectConnectionPage
-    cancelHref={(p, s) => resolvers.create.configure.index({ ...p, ...s })}
-    apiProviderHref={resolvers.create.configure.addStep.apiProvider.upload}
-    connectionHref={(connection, p, s) =>
-      resolvers.create.configure.addStep.connection.selectAction({
-        connection,
-        ...p,
-        ...s,
-      })
-    }
-    filterHref={resolvers.create.configure.addStep.basicFilter}
-    mapperHref={resolvers.create.configure.addStep.dataMapper}
-    templateHref={(step, params, state) =>
-      resolvers.create.configure.addStep.template({
-        step,
-        ...params,
-        ...state,
-      })
-    }
-    stepHref={(step, params, state) =>
-      resolvers.create.configure.addStep.step({
-        step,
-        ...params,
-        ...state,
-      })
-    }
-    sidebar={({ steps, activeIndex }) => (
-      <IntegrationEditorSidebar
-        steps={steps}
-        addAtIndex={activeIndex}
-        addI18nTitle={`${activeIndex + 1}. Start`}
-        addI18nTooltip={'Start'}
-        addI18nDescription={'Choose a connection'}
-      />
-    )}
-  />
-);
-
-const addStepSelectActionPage = (
-  <SelectActionPage
-    cancelHref={(p, s) => resolvers.create.configure.index({ ...p, ...s })}
-    sidebar={({ connection, steps, activeIndex }) => (
-      <IntegrationEditorSidebar
-        steps={steps}
-        addAtIndex={activeIndex}
-        addIcon={
-          <img
-            alt={'Step icon'}
-            src={getConnectionIcon(process.env.PUBLIC_URL, connection)}
-            height={24}
-            width={24}
-          />
-        }
-        addI18nTitle={`${activeIndex + 1}. ${connection.connector!.name}`}
-        addI18nTooltip={`${activeIndex + 1}. ${connection.name}`}
-        addI18nDescription={'Choose an action'}
-      />
-    )}
-    selectHref={(actionId, p, s) =>
-      resolvers.create.configure.addStep.connection.configureAction({
-        actionId,
-        ...p,
-        ...s,
-      })
-    }
-  />
-);
-
-const addStepConfigureActionPage = (
-  <ConfigureActionPage
-    backHref={(p, s) =>
-      resolvers.create.configure.addStep.connection.selectAction({ ...p, ...s })
-    }
-    cancelHref={(p, s) => resolvers.create.configure.index({ ...p, ...s })}
-    mode={'adding'}
-    nextStepHref={(p, s) =>
-      resolvers.create.configure.addStep.connection.configureAction({
-        ...p,
-        ...s,
-      })
-    }
-    sidebar={({ connection, steps, activeIndex }) => (
-      <IntegrationEditorSidebar
-        steps={steps}
-        addAtIndex={activeIndex}
-        addIcon={
-          <img
-            alt={'Step icon'}
-            src={getConnectionIcon(process.env.PUBLIC_URL, connection)}
-            height={24}
-            width={24}
-          />
-        }
-        addI18nTitle={`${activeIndex + 1}. ${connection.connector!.name}`}
-        addI18nTooltip={`${activeIndex + 1}. ${connection.name}`}
-        addI18nDescription={'Configure the action'}
-      />
-    )}
-    postConfigureHref={(integration, params) =>
-      resolvers.create.configure.index({
-        integration,
-        ...params,
-      })
-    }
-  />
-);
-
-const addStepConfigureStepPage = (
-  <ConfigureStepPage
-    cancelHref={(p, s) => resolvers.create.configure.index({ ...p, ...s })}
-    mode={'adding'}
-    sidebar={({ step, steps, activeIndex }) => (
-      <IntegrationEditorSidebar
-        steps={steps}
-        addAtIndex={activeIndex}
-        addIcon={
-          <img
-            alt={'Step icon'}
-            src={getStepIcon(process.env.PUBLIC_URL, step)}
-            height={24}
-            width={24}
-          />
-        }
-        addI18nTitle={`${activeIndex + 1}. ${step.name}`}
-        addI18nTooltip={`${activeIndex + 1}. ${step.description}`}
-        addI18nDescription={'Configure the action'}
-      />
-    )}
-    postConfigureHref={(integration, params) =>
-      resolvers.create.configure.index({
-        integration,
-        ...params,
-      })
-    }
-  />
-);
-
-const editStepSelectActionPage = (
-  <SelectActionPage
-    cancelHref={(p, s) => resolvers.create.configure.index({ ...p, ...s })}
-    sidebar={({ steps, activeIndex }) => (
-      <IntegrationEditorSidebar steps={steps} activeIndex={activeIndex} />
-    )}
-    selectHref={(actionId, p, s) =>
-      resolvers.create.configure.editStep.connection.configureAction({
-        actionId,
-        ...p,
-        ...s,
-      })
-    }
-  />
-);
-
-const editStepConfigureActionPage = (
-  <ConfigureActionPage
-    backHref={(p, s) =>
-      resolvers.create.configure.editStep.connection.selectAction({
-        ...p,
-        ...s,
-      })
-    }
-    cancelHref={(p, s) => resolvers.create.configure.index({ ...p, ...s })}
-    mode={'editing'}
-    nextStepHref={(p, s) =>
-      resolvers.create.configure.editStep.connection.configureAction({
-        ...p,
-        ...s,
-      })
-    }
-    sidebar={({ steps, activeIndex }) => (
-      <IntegrationEditorSidebar steps={steps} activeIndex={activeIndex} />
-    )}
-    postConfigureHref={(integration, params) =>
-      resolvers.create.configure.index({
-        integration,
-        ...params,
-      })
-    }
-  />
-);
-
-const editStepConfigureStepPage = (
-  <ConfigureStepPage
-    cancelHref={(p, s) => resolvers.create.configure.index({ ...p, ...s })}
-    mode={'editing'}
-    sidebar={({ steps, activeIndex }) => (
-      <IntegrationEditorSidebar steps={steps} activeIndex={activeIndex} />
-    )}
-    postConfigureHref={(integration, params) =>
-      resolvers.create.configure.index({
-        integration,
-        ...params,
-      })
-    }
-  />
-);
-
-const addTemplateStepPage = (
-  <TemplateStepPage
-    mode={'adding'}
-    cancelHref={(p, s) => resolvers.create.configure.index({ ...p, ...s })}
-    sidebar={({ steps, activeIndex }) => (
-      <IntegrationEditorSidebar steps={steps} activeIndex={activeIndex} />
-    )}
-    postConfigureHref={(integration, params) =>
-      resolvers.create.configure.index({
-        integration,
-        ...params,
-      })
-    }
-  />
-);
-
-const editTemplateStepPage = (
-  <TemplateStepPage
-    mode={'editing'}
-    cancelHref={(p, s) => resolvers.create.configure.index({ ...p, ...s })}
-    sidebar={({ steps, activeIndex }) => (
-      <IntegrationEditorSidebar steps={steps} activeIndex={activeIndex} />
-    )}
-    postConfigureHref={(integration, params) =>
-      resolvers.create.configure.index({
-        integration,
-        ...params,
-      })
-    }
-  />
-);
-
-const TODO: React.FunctionComponent = () => <>TODO</>;
-
 /**
  * Entry point for the integration creator app. This is shown when an user clicks
  * the "Create integration" button somewhere in the app.
@@ -743,44 +91,16 @@ export const IntegrationCreatorApp: React.FunctionComponent = () => {
         {/* start step */}
         <Route path={routes.create.start.selectStep}>
           <EditorApp
-            selectStepPath={routes.create.start.selectStep}
-            selectStepChildren={startStepSelectConnectionPage}
-            endpointEditor={{
-              selectActionPath: routes.create.start.connection.selectAction,
-              selectActionChildren: startStepSelectActionPage,
-              configureActionPath:
-                routes.create.start.connection.configureAction,
-              configureActionChildren: startStepConfigureActionPage,
-              describeDataPath: routes.create.start.connection.describeData,
-              describeDataChildren: TODO,
-            }}
-            apiProvider={{
-              uploadPath: routes.create.start.apiProvider.upload,
-              uploadChildren: <UploadPage />,
-              reviewPath: routes.create.start.apiProvider.review,
-              reviewChildren: <ReviewPage />,
-              editPath: routes.create.start.apiProvider.edit,
-              editChildren: <EditPage />,
-            }}
-            template={{
-              templatePath: routes.create.start.template,
-              templateChildren: addTemplateStepPage,
-            }}
-            dataMapper={{
-              mapperPath: routes.create.start.dataMapper,
-              mapperChildren: TODO,
-            }}
-            basicFilter={{
-              filterPath: routes.create.start.basicFilter,
-              filterChildren: TODO,
-            }}
-            step={{
-              configurePath: routes.create.start.step,
-              configureChildren: startStepConfigureStepPage,
-            }}
-            extension={{
-              configurePath: routes.create.start.extension,
-              configureChildren: TODO,
+            mode={'adding'}
+            appStepRoutes={routes.create.start}
+            appResolvers={resolvers.create.start}
+            cancelHref={resolvers.list}
+            postConfigureHref={(integration, params) => {
+              return resolvers.create.finish.selectStep({
+                integration,
+                ...params,
+                position: '1',
+              });
             }}
           />
         </Route>
@@ -788,45 +108,16 @@ export const IntegrationCreatorApp: React.FunctionComponent = () => {
         {/* finish step */}
         <Route path={routes.create.finish.selectStep}>
           <EditorApp
-            selectStepPath={routes.create.finish.selectStep}
-            selectStepChildren={finishStepSelectConnectionPage}
-            endpointEditor={{
-              selectActionPath: routes.create.finish.connection.selectAction,
-              selectActionChildren: finishStepSelectActionPage,
-              configureActionPath:
-                routes.create.finish.connection.configureAction,
-              configureActionChildren: finishStepConfigureActionPage,
-              describeDataPath: routes.create.finish.connection.describeData,
-              describeDataChildren: TODO,
-            }}
-            apiProvider={{
-              uploadPath: routes.create.finish.apiProvider.upload,
-              uploadChildren: <UploadPage />,
-              reviewPath: routes.create.finish.apiProvider.review,
-              reviewChildren: <ReviewPage />,
-              editPath: routes.create.finish.apiProvider.edit,
-              editChildren: <EditPage />,
-            }}
-            template={{
-              templatePath: routes.create.finish.template,
-              templateChildren: addTemplateStepPage,
-            }}
-            dataMapper={{
-              mapperPath: routes.create.finish.dataMapper,
-              mapperChildren: TODO,
-            }}
-            basicFilter={{
-              filterPath: routes.create.finish.basicFilter,
-              filterChildren: TODO,
-            }}
-            step={{
-              configurePath: routes.create.finish.step,
-              configureChildren: finishStepConfigureStepPage,
-            }}
-            extension={{
-              configurePath: routes.create.finish.extension,
-              configureChildren: TODO,
-            }}
+            mode={'adding'}
+            appStepRoutes={routes.create.finish}
+            appResolvers={resolvers.create.finish}
+            cancelHref={resolvers.list}
+            postConfigureHref={(integration, params) =>
+              resolvers.create.configure.index({
+                integration,
+                ...params,
+              })
+            }
           />
         </Route>
 
@@ -839,92 +130,32 @@ export const IntegrationCreatorApp: React.FunctionComponent = () => {
         {/* add step */}
         <Route path={routes.create.configure.addStep.selectStep}>
           <EditorApp
-            selectStepPath={routes.create.configure.addStep.selectStep}
-            selectStepChildren={addStepSelectConnectionPage}
-            endpointEditor={{
-              selectActionPath:
-                routes.create.configure.addStep.connection.selectAction,
-              selectActionChildren: addStepSelectActionPage,
-              configureActionPath:
-                routes.create.configure.addStep.connection.configureAction,
-              configureActionChildren: addStepConfigureActionPage,
-              describeDataPath:
-                routes.create.configure.addStep.connection.describeData,
-              describeDataChildren: TODO,
-            }}
-            apiProvider={{
-              uploadPath: routes.create.configure.addStep.apiProvider.upload,
-              uploadChildren: <UploadPage />,
-              reviewPath: routes.create.configure.addStep.apiProvider.review,
-              reviewChildren: <ReviewPage />,
-              editPath: routes.create.configure.addStep.apiProvider.edit,
-              editChildren: <EditPage />,
-            }}
-            template={{
-              templatePath: routes.create.configure.addStep.template,
-              templateChildren: addTemplateStepPage,
-            }}
-            dataMapper={{
-              mapperPath: routes.create.configure.addStep.dataMapper,
-              mapperChildren: TODO,
-            }}
-            basicFilter={{
-              filterPath: routes.create.configure.addStep.basicFilter,
-              filterChildren: TODO,
-            }}
-            step={{
-              configurePath: routes.create.configure.addStep.step,
-              configureChildren: addStepConfigureStepPage,
-            }}
-            extension={{
-              configurePath: routes.create.configure.addStep.extension,
-              configureChildren: TODO,
-            }}
+            mode={'adding'}
+            appStepRoutes={routes.create.configure.addStep}
+            appResolvers={resolvers.create.configure.addStep}
+            cancelHref={resolvers.list}
+            postConfigureHref={(integration, params) =>
+              resolvers.create.configure.index({
+                integration,
+                ...params,
+              })
+            }
           />
         </Route>
 
         {/* edit step */}
         <Route path={routes.create.configure.editStep.selectStep}>
           <EditorApp
-            endpointEditor={{
-              selectActionPath:
-                routes.create.configure.editStep.connection.selectAction,
-              selectActionChildren: editStepSelectActionPage,
-              configureActionPath:
-                routes.create.configure.editStep.connection.configureAction,
-              configureActionChildren: editStepConfigureActionPage,
-              describeDataPath:
-                routes.create.configure.editStep.connection.describeData,
-              describeDataChildren: TODO,
-            }}
-            apiProvider={{
-              uploadPath: routes.create.configure.editStep.apiProvider.upload,
-              uploadChildren: <UploadPage />,
-              reviewPath: routes.create.configure.editStep.apiProvider.review,
-              reviewChildren: <ReviewPage />,
-              editPath: routes.create.configure.editStep.apiProvider.edit,
-              editChildren: <EditPage />,
-            }}
-            template={{
-              templatePath: routes.create.configure.editStep.template,
-              templateChildren: editTemplateStepPage,
-            }}
-            dataMapper={{
-              mapperPath: routes.create.configure.editStep.dataMapper,
-              mapperChildren: TODO,
-            }}
-            basicFilter={{
-              filterPath: routes.create.configure.editStep.basicFilter,
-              filterChildren: TODO,
-            }}
-            step={{
-              configurePath: routes.create.configure.editStep.step,
-              configureChildren: editStepConfigureStepPage,
-            }}
-            extension={{
-              configurePath: routes.create.configure.editStep.extension,
-              configureChildren: TODO,
-            }}
+            mode={'editing'}
+            appStepRoutes={routes.create.configure.editStep}
+            appResolvers={resolvers.create.configure.editStep}
+            cancelHref={resolvers.list}
+            postConfigureHref={(integration, params) =>
+              resolvers.create.configure.index({
+                integration,
+                ...params,
+              })
+            }
           />
         </Route>
