+++ app/ui-react/syndesis/src/modules/integrations/components/index.ts
@@ -3,6 +3,5 @@ export * from './IntegrationCreatorBreadcrumbs';
 export * from './IntegrationDetailHeader';
 export * from './IntegrationDetailSteps';
 export * from './IntegrationEditorBreadcrumbs';
-export * from './IntegrationEditorSidebar';
 export * from './IntegrationEditorStepAdder';
 export * from './WithIntegrationActions';
