+++ app/ui-react/syndesis/src/modules/integrations/routes.ts
@@ -1,45 +1,6 @@
 /* tslint:disable:object-literal-sort-keys */
 import { include } from 'named-urls';
-
-export const stepRoutes = {
-  // step 1
-  selectStep: '',
-  // if selected step is api provider
-  apiProvider: include('api-provider', {
-    upload: '',
-    review: 'review',
-    edit: 'edit',
-  }),
-  // if selected step kind is data mapper
-  dataMapper: 'mapper',
-  // if selected step kind is basic filter
-  basicFilter: 'filter',
-  // if selected step kind is template
-  template: 'template',
-  // if selected step kind is step
-  step: 'step',
-  // if selected step kind is step
-  extension: 'extension',
-  // if selected step kind is endpoint
-  connection: include('connection/:connectionId', {
-    selectAction: '',
-    configureAction: ':actionId/:step',
-    // if 'any' data shape
-    describeData: 'describe-data/:position/:direction(input|output)',
-  }),
-};
-
-/**
- * Both the integration creator and editor share the same routes when the creator
- * reaches the third step in the wizard. This object is to keep them DRY.
- */
-const editorRoutes = include(':flowId', {
-  index: 'add-step',
-  addStep: include(':position/add', stepRoutes),
-  editStep: include(':position/edit', stepRoutes),
-  saveAndPublish: 'save',
-  root: '',
-});
+import { editorRoutes, stepRoutes } from './components/editor/interfaces';
 
 export default include('/integrations', {
   list: '',
